+++ src/collective/documentgenerator/browser/generation_view.py
@@ -218,6 +218,7 @@ class DocumentGenerationView(BrowserView):
             forceOoCall=True,
             optimalColumnWidths=optimalColumnWidths,
             stylesMapping=stylesMapping,
+            renamePageStyles=True,
             **kwargs
         )